requirements.txt
# FILE: requirements.txt
Flask>=3.0.0
Flask-SQLAlchemy>=3.1.1
python-dotenv>=1.0.1
google-genai>=0.8.0
pandas>=2.2.0
scikit-learn>=1.4.0
numpy>=1.26.0
scipy>=1.12.0
joblib>=1.3.0
